+++ test/twitter.test.ts
@@ -2,14 +2,21 @@ import { test, expect } from "vitest"
 import { getBrowserContext, getFollowerCount, destroyBrowser } from "../src"
 
 test("twitter", async () => {
-  const count = await getFollowerCount({
-    type: "twitter",
-    username: "cristiano",
-    browserContext: await getBrowserContext(),
-  })
+  const [countByBrowser, countByApi] = await Promise.all([
+    getFollowerCount({
+      type: "twitter",
+      username: "cristiano",
+      browserContext: await getBrowserContext(),
+    }),
+    getFollowerCount({
+      type: "twitter",
+      username: "cristiano",
+    }),
+  ])
 
   await destroyBrowser()
 
-  console.log(count)
-  expect(count).toBeGreaterThan(0)
+  console.log(countByBrowser, countByApi)
+  expect(countByBrowser).toBeGreaterThan(0)
+  expect(countByApi).toBeGreaterThan(0)
 })