+++ CHANGELOG.md
@@ -1,6 +1,6 @@
 ## Unreleased
 
-No unreleased changes.
+- twitter: Now you don't need headless chrome to fetch the follower count, it uses the twitter follow button api if a `browserContext` is not provided.
 
 ## 0.2.2